Move vector into primitives
