+++ library/Rattletrap/AttributeValue/RigidBodyState.hs
@@ -1,7 +1,7 @@
 module Rattletrap.AttributeValue.RigidBodyState where
 
 import Rattletrap.Primitive.CompressedWordVector
-import Rattletrap.Vector
+import Rattletrap.Primitive.Vector
 
 import qualified Data.Binary.Bits.Get as BinaryBit
 import qualified Data.Binary.Bits.Put as BinaryBit
